chore: initial chart library removal
