+++ src/vaults/vaults.service.ts
@@ -311,17 +311,6 @@ export class VaultsService {
     };
   }
 
-  async loadVaultChartData(
-    vaultAddr: string,
-    timeframe: ChartTimeFrame,
-    chain: Chain,
-  ): Promise<HistoricVaultSnapshotModel[]> {
-    const vaultBlobId = HistoricVaultSnapshotModel.formBlobId(vaultAddr, chain.network);
-    const dataKey = toChartDataKey(HistoricVaultSnapshotModel.NAMESPACE, vaultBlobId, timeframe);
-
-    return queryVaultCharts(dataKey);
-  }
-
   async getVaultChartDataByTimestamps(
     vaultAddr: string,
     network: Network,
